* GWMetadata/MDFinder/MDFinder.[h|m]         * GWMetadata/Preferences/StartAppWin.[h|m]         * GWorkspace/Dialogs/StartAppWin.[h|m]         * Inspector/ContentViewers/SoundViewer/SoundViewer.m         * Recycler/Dialogs/StartAppWin.[h|m]         NSProgressIndicator uses doubles, so GWorkspace         should use doubles too
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@37264 72102866-910b-0410-8b05-ffd578937521

+++ GWMetadata/MDFinder/MDFinder.h
@@ -103,9 +103,9 @@
 - (void)showWindowWithTitle:(NSString *)title
                     appName:(NSString *)appname
                   operation:(NSString *)operation              
-               maxProgValue:(float)maxvalue;
+               maxProgValue:(double)maxvalue;
 
-- (void)updateProgressBy:(float)incr;
+- (void)updateProgressBy:(double)incr;
 
 - (NSWindow *)win;
