+++ Recycler/Dialogs/StartAppWin.h
@@ -37,9 +37,9 @@
                  
 - (void)showWindowWithTitle:(NSString *)title
                     appName:(NSString *)appname
-               maxProgValue:(float)maxvalue;
+               maxProgValue:(double)maxvalue;
 
-- (void)updateProgressBy:(float)incr;
+- (void)updateProgressBy:(double)incr;
 
 - (NSWindow *)win;
